+++ packages/expo-image-manipulator/android/src/main/java/expo/modules/imagemanipulator/ImageRef.kt
@@ -5,6 +5,8 @@ import expo.modules.kotlin.RuntimeContext
 import expo.modules.kotlin.sharedobjects.SharedRef
 
 class ImageRef(bitmap: Bitmap, runtimeContext: RuntimeContext) : SharedRef<Bitmap>(bitmap, runtimeContext) {
+  override val nativeRefType: String = "image"
+
   override fun getAdditionalMemoryPressure(): Int {
     return ref.allocationByteCount
   }