+++ packages/expo-image-manipulator/ios/ImageRef.swift
@@ -6,6 +6,12 @@ import ExpoModulesCore
  Represents a shared reference to the `UIImage` instance.
  */
 internal final class ImageRef: SharedRef<UIImage> {
+  override var nativeRefType: String {
+    get {
+      "image"
+    }
+  }
+  
   override func getAdditionalMemoryPressure() -> Int {
     guard let cgImage = ref.cgImage else {
       return 0