+++ packages/expo-image/ios/Image.swift
@@ -3,6 +3,12 @@
 import ExpoModulesCore
 
 internal final class Image: SharedRef<UIImage> {
+  override var nativeRefType: String {
+    get {
+      "image"
+    }
+  }
+  
   override func getAdditionalMemoryPressure() -> Int {
     guard let cgImage = ref.cgImage else {
       return 0