+++ packages/expo-modules-core/android/src/androidTest/java/expo/modules/kotlin/jni/SharedRefTest.kt
@@ -4,7 +4,9 @@ import com.google.common.truth.Truth
 import expo.modules.kotlin.sharedobjects.SharedRef
 import org.junit.Test
 
-class SharedString(value: String) : SharedRef<String>(value)
+class SharedString(value: String) : SharedRef<String>(value) {
+  override val nativeRefType: String = "string"
+}
 
 class SharedRefTest {
   @Test
@@ -42,5 +44,20 @@ class SharedRefTest {
       """.trimIndent()
     ).getString()
     Truth.assertThat(string).isEqualTo("shared string")
+  }
+
+  @Test
+  fun has_native_ref_type() = withJSIInterop(
+    inlineModule {
+      Name("TestModule")
+      Class("SharedString", SharedString::class) {
+        Constructor {
+          SharedString("shared string")
+        }
+      }
+    }
+  ) {
+    val nativeRefType = evaluateScript("(new expo.modules.TestModule.SharedString()).nativeRefType").getString()
+    Truth.assertThat(nativeRefType).isEqualTo("string")
   }
 }