+++ packages/expo-modules-core/ios/Core/Classes/ClassDefinition.swift
@@ -32,12 +32,19 @@ public final class ClassDefinition: ObjectDefinition {
     self.name = name
     self.constructor = elements.first(where: isConstructor) as? AnySyncFunctionDefinition
     self.associatedType = ~AssociatedObject.self
-    self.isSharedRef = AssociatedObject.self is SharedRefAssociatedObject.Type
+    self.isSharedRef = AssociatedObject.self is AnySharedRef.Type
 
     // Constructors can't be passed down to the object definition
     // as we shouldn't override the default `<Class>.prototype.constructor`.
-    let elementsWithoutConstructors = elements.filter({ !isConstructor($0) })
-
+    var elementsWithoutConstructors = elements.filter({ !isConstructor($0) })
+    if (isSharedRef) {
+      elementsWithoutConstructors.append(
+        PropertyDefinition<AssociatedObject>(name: "nativeRefType", getter: { (owner: AssociatedObject) -> String in
+          return (owner as! AnySharedRef).nativeRefType
+        })
+      )
+    }
+    
     super.init(definitions: elementsWithoutConstructors)
   }
 
@@ -108,12 +115,6 @@ extension JavaScriptObject: ClassAssociatedObject, AnyArgument, AnyJavaScriptVal
   }
 }
 extension SharedObject: ClassAssociatedObject {}
-
-/**
- A protocol that allows us to detect `SharedRef`'s.
- */
-internal protocol SharedRefAssociatedObject {}
-extension SharedRef : SharedRefAssociatedObject {}
 
 // MARK: - Privates